회원가입 페이지 및 분석보고서 페이지

+++ src/main/java/com/example/etProject/entity/ProducersEntity.java
@@ -42,7 +42,7 @@ public class ProducersEntity {
     private Double locationY;
     @Column(name="INSTALLED_CAPACITY")
     @ColumnDefault("-1")
-    private Double installedCapacity;
+    private int installedCapacity;
     @Column(name="IS_PRODUCE")
     private Boolean isProduce;
